Initial Commit
Random number generated upon page load.
Game will ask for a number if input is NaN.
Game lets user know if high or low.
Game asks to replay if user has won.

KNOWN ERRORS:
- Number is not reset if starting a new game without a page load. (Easy
fix. Add a line to run the randomizer after the confirm is true.)
- If game is won, game will show an additional congratulations message
for every time a wrong number was guessed. (Still looking for a source
to this problem...)